feat: Added script to fix errors in osm files
This script has been created to remove entities that are source of errors in
a osm file. Most of the errors were caused because of bugs in the Open Studio
plugin for Sketchup. Most recent version of Open Studio have been improved,
so no errors are found when loading the model.